// src/modules/ehr.analytics/ehr.domain.models/ehr.record.types.ts
export enum EHRRecordTypes {
    Birthdate = 'Birthdate',
    Gender = 'Gender',
    Race = 'Race',
    Ethnicity = 'Ethnicity',
    Diabetic = 'Diabetic',
    Smoker = 'Smoker',
    BloodGroup = 'BloodGroup',
    HealthSystem = 'Health-System',
    BodyHeight = 'Body-Height',
    BodyWeight = 'Body-Weight',
    BodyTemperature = 'Body-Temperature',
    BloodPressure = 'Blood-Pressure',
    BloodGlucose = 'Blood-Glucose',
    Cholesterol_Total = 'Cholesterol-Total',
    Cholesterol_HDL = 'Cholesterol-HDL',
    Cholesterol_LDL = 'Cholesterol-LDL',
    Cholesterol_TriglycerideLevel = 'Cholesterol-Triglyceride-Level',
    Cholesterol_Ratio = 'Cholesterol-Ratio',
    Cholesterol_A1CLevel = 'Cholesterol-A1C-Level',
    BloodOxygenSaturation = 'Blood-Oxygen-Saturation',
    Pulse = 'Pulse',
    Symptom = 'Symptom',
    SymptomMood = 'Symptom-Mood',
    SymptomFeeling = 'Symptom-Feeling',
    SymptomEnergyLevels = 'Symptom-Energy-Levels',
    MedicationIsTaken = 'Medication-IsTaken ',
    MedicationIsMissed = 'Medication-IsMissed ',
    Feeling = 'Feeling',
    Mood = 'Mood',
    EnergyLevel = 'Energy-Level',
    LabRecord = 'Lab-Record',
    PhysicalActivity = 'Physical-Activity',
    PhysicalActivityStepCounts = 'Physical-Activity-Step-Counts',
    PhysicalActivityStand = 'Physical-Activity-Stand',
    PhysicalActivityExercise = 'Physical-Activity-Exercise',
    MentalWellBeing = 'Mental-Well-Being',
    MentalWellBeingSleep = 'Mental-Well-Being-Sleep',
    MentalWellBeingMeditation = 'Mental-Well-Being-Meditation',
    Nutrition = 'Nutrition',
    NutritionFruit = 'Nutrition-Fruit',
    NutritionVegetables = 'Nutrition-Vegetables',
    NutritionSugaryDrinks = 'Nutrition-SugaryDrinks',
    NutritionSalt = 'Nutrition-Salt',
    NutritionSeaFood = 'Nutrition-SeaFood',
    NutritionGrains = 'Nutrition-Grains',
    NutritionProtein = 'Nutrition-Protein',
    Other = 'Other'
}

export enum DataTypes {
    Integer = 'Integer',
    Boolean = 'Boolean',
    Float = 'Float',
    String = 'String'
}

export enum EHRModels {
    'Body-Height' = 'EHRVitalData',
    'Body-Weight' = 'EHRVitalData',
    'Body-Temperature' = 'EHRVitalData',
    'Blood-Pressure' = 'EHRVitalData',
    'Blood-Glucose' = 'EHRVitalData',
    'Blood-Oxygen-Saturation' = 'EHRVitalData',
    'Pulse' = 'EHRVitalData',
    'Cholesterol-Total' = 'EHRLabData',
    'Cholesterol-HDL' = 'EHRLabData',
    'Cholesterol-LDL' = 'EHRLabData',
    'Cholesterol-Triglyceride-Level' = 'EHRLabData',
    'Cholesterol-Ratio' = 'EHRLabData',
    'Cholesterol-A1C-Level' = 'EHRLabData',
    'Symptom' = 'EHRSymptomData',
    'Medication-IsTaken ' = 'EHRMedicationData',
    'Medication-IsMissed ' = 'EHRMedicationData',
    'Lab-Record' = 'EHRLabData',
    'Physical-Activity' = 'EHRPhysicalActivityData',
    'Mental-Well-Being' = 'EHRMentalWellBeingData',
    'Nutrition' = 'EHRNutritionData',
    'Nutrition-Fruit' = 'EHRNutritionData',
    'Nutrition-Vegetables' = 'EHRNutritionData',
    'Nutrition-SugaryDrinks' = 'EHRNutritionData',
    'Nutrition-Salt' = 'EHRNutritionData',
    'Nutrition-SeaFood' = 'EHRNutritionData',
    'Nutrition-Grains' = 'EHRNutritionData',
    'Nutrition-Protein' = 'EHRNutritionData'
}